requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
colorama==0.4.6
fastapi==0.136.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
mangum==0.21.0
psycopg2-binary==2.9.11
pydantic==2.13.2
pydantic_core==2.46.2
python-dotenv==1.2.2
SQLAlchemy==2.0.49
sqlmodel==0.0.38
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uuid6==2025.0.1
uvicorn==0.44.0
